ORG: Restructuring metadata related objects into their own package
