+++ skbio/metadata/_repr.py
